chore: Move writer-related logic to "writers" module (#3385)
